holy crap... a lot... i need to commit more often. imgui back in, tons of mesh and model stuff. modified imgui-go from inkyblackness

+++ mesh/cube.go
@@ -75,7 +75,7 @@ var cubePositions = []float32{
 	0.5, 0.5, 0.5,
 }
 
-var cubeIndices = []uint16{
+var cubeIndices = []uint32{
 	// front
 	0, 1, 2,
 	2, 3, 0,
